+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Release 5.6.6:
  - OpenID for Verifiable Presentations:
    - Fix applying presentation exchange filters to credentials (`array` and `object` filters)
+ - OpenID for Verifiable Credential Issuance:
+   - On issued SD-JWT VC do not validate subject but the confirmation claim
 
 Release 5.6.5:
  - OpenID for Verifiable Presentations: